+++ R/zzz.R
@@ -1,2 +1,9 @@
 #' @importFrom utils getFromNamespace
 NULL
+
+
+
+backports = list(
+  base = c("anyNA", "dir.exists", "endsWith", "file.info", "file.mode", "file.mtime", "file.size", "get0", "lengths", "startsWith", "strrep", "trimws"),
+  utils = "hasName"
+)