+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Added examples.
 * Added possibility to set `method = "radix"` in `sort()` and `order()` for R versions
-  prior to 3.3.0 with a fallback to shell sort.
+  prior to 3.3.0 with a fallback to shell sort with `LC_COLLATE="C"`.
 
 # backports 1.0.4